requirements.txt
dnspython>=2.7.0,<3.0.0
fastapi>=0.115.0,<1.0.0
pymongo>=4.11.0,<5.0.0
python-dotenv>=1.1.0,<2.0.0
python-telegram-bot>=22.7,<23.0
uvicorn[standard]>=0.34.0,<1.0.0
